+++ testnet/stacks-node/src/nakamoto_node/stackerdb_listener.rs
@@ -312,7 +312,7 @@ impl StackerDBListener {
                 "slot_ids" => ?slot_ids,
             );
 
-            for ((_, message), slot_id) in messages.into_iter().zip(slot_ids) {
+            for (message, slot_id) in messages.into_iter().zip(slot_ids) {
                 let Some(signer_entry) = &self.signer_entries.get(&slot_id) else {
                     return Err(NakamotoNodeError::SignerSignatureError(
                         "Signer entry not found".into(),
@@ -498,7 +498,15 @@ impl StackerDBListener {
                         debug!("Received mock message. Ignoring.");
                     }
                     SignerMessageV0::StateMachineUpdate(update) => {
-                        self.update_global_state_evaluator(&signer_pubkey, update);
+                        let Some(pubkey) = update.recover_signer_pk() else {
+                            warn!("Received a state machine update with a non-recoverable signature. Ignoring.";
+                                "update" => ?update,
+                                "slot_pubkey" => ?signer_pubkey,
+                                "slot_id" => slot_id,
+                            );
+                            continue;
+                        };
+                        self.update_global_state_evaluator(&pubkey, update);
                     }
                 };
             }