+++ stacks-signer/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to the versioning scheme outlined in the [README.md](README.md).
 
+## [Unreleased]
+
+### Added
+- Introduced `capitulate_tenure_timeout`: the duration (in seconds) for the signer to wait between updating the local state machine viewpoint and capitulating to other signers' miner views.
+
+### Changed
+- Updated signer block proposal evaluation to utilize the global signer state, enhancing consistency and reliability in consensus decisions.
+
 ## [3.1.0.0.10.0]
 
 ### Added